+++ main.py
@@ -4,7 +4,7 @@
 # Created on: 2018-06-02
 # License: GPL v.3 https://www.gnu.org/copyleft/gpl.html
 
-import menu_helper
+import waipu
 
 if __name__ == '__main__':
-    menu_helper.run()
+    waipu.run()
